requirements.txt
Pillow
ips-py
pyinstaller